+++ src/test/java/BasicStructure.java
@@ -1,3 +1,5 @@
+package test.java;
+
 public class BasicStructure {
     String Sun;
     String Moon;
