+++ src/test/java/FeatureSun.java
@@ -1,0 +1,11 @@
+package test.java;
+
+public class FeatureSun {
+
+        String Color;
+        String tem;
+        String Radius;
+
+
+    }
+
