--- src/test/java/SolarS.java
@@ -1,25 +1,0 @@
-public class SolarS {
-    public static void main(String[] args) {
-
-       BasicStructure BB=new BasicStructure();
-       BB.Planets="2";
-       BB.Moon="2";
-       BB.Stars="10";
-       BB.Sun="1";
-        System.out.println("Solar is ready");
-
-
-
-
-
-    }
-
-
-
-
-
-
-
-
-
-}
