+++ src/test/java/SolarSys.java
@@ -1,0 +1,20 @@
+package test.java;
+
+public class SolarSys {
+        public static void main(String[] args) {
+
+            BasicStructure BB=new BasicStructure();
+            BB.Planets="2";
+            BB.Moon="2";
+            BB.Stars="10";
+            BB.Sun="1";
+            System.out.println("Solar is ready");
+
+            FeatureSun Xros=new FeatureSun();
+            Xros.tem="324325kd";
+            Xros.Color="green";
+            Xros.Radius="6458593754759";
+
+
+        }
+}
